LPD-70615 Removed Get App Component
