+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/components/ProductPurchase/Header/index.tsx
@@ -7,7 +7,7 @@ import {ComponentProps} from 'react';
 
 import i18n from '../../../i18n';
 import AccountEmailInfo from '../../../pages/CustomerDashboard/pages/Apps/App/Licenses/CreateLicense/AccountInfo';
-import {ProductCardRevamp as ProductCard} from '../../../pages/GetApp/components/ProductCard/ProductCard';
+import {ProductCardRevamp as ProductCard} from '../../../pages/ProductPurchase/components/ProductCard/ProductCard';
 import {normalizeURLProtocol} from '../../../utils/string';
 
 type ProductPurchaseHeaderAccountProps = {
