+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/CustomerDashboard/components/LicenseDetailsModalHeader.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {getIconUrl} from '../../GetApp/utils/getIcon';
+import {getIconUrl} from '../../../utils/getIcon';
 import AccountEmailInfo from '../pages/Apps/App/Licenses/CreateLicense/AccountInfo';
 import OrderDetailsHeader from './OrderDetailsHeader';
