--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/GetApp/enums/GetAppStepTypes.ts
@@ -1,11 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-export enum GetAppStepTypes {
-	ACCOUNT = 'account',
-	LICENSES = 'licenses',
-	PAYMENT = 'payment',
-	PROJECT = 'project',
-}
