--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/GetApp/utils/getReplaceCurrentURL.ts
@@ -1,23 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {Liferay} from '../../../liferay/liferay';
-
-export default function getReplaceCurrentURL(
-	currentURL: string,
-	nextURL: string,
-	orderId?: string
-) {
-	let newURL = `${Liferay.ThemeDisplay.getLayoutURL().replace(
-		`/${currentURL}`,
-		''
-	)}/${nextURL}`;
-
-	if (orderId) {
-		newURL = `${newURL}?orderId=${orderId}`;
-	}
-
-	return newURL;
-}
