--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/GetApp/utils/postCartByPaymentMethod.ts
@@ -1,13 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {postCartByChannelId} from '../../../utils/api';
-
-export async function postCartByPaymentMethod(cart: {}, channelId: number) {
-	return await postCartByChannelId({
-		cartBody: cart,
-		channelId,
-	});
-}
