+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/OAuth2Authorize/OAuth2AuthorizeOutlet.tsx
@@ -10,7 +10,7 @@ import {Outlet, useOutletContext} from 'react-router-dom';
 import {useMarketplaceContext} from '../../context/MarketplaceContext';
 import zodSchema, {z, zodResolver} from '../../schema/zod';
 import {ConsoleUserProject} from '../../services/oauth/types';
-import useGetResourceInfo from '../GetApp/hooks/useGetResourceInfo';
+import useGetResourceInfo from '../ProductPurchase/hooks/useGetResourceInfo';
 import useAccounts from '../ProductPurchase/hooks/useAccounts';
 
 type Schema = z.infer<typeof zodSchema.installProductSchema>;
