+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/OAuth2Authorize/pages/EnvironmentSelection.tsx
@@ -12,7 +12,7 @@ import i18n from '../../../i18n';
 import ContactSupport from '../../CustomerDashboard/pages/Apps/App/CloudProvisioning/components/ContactSupport';
 import SelectedProjectBanner from '../../CustomerDashboard/pages/Apps/App/CloudProvisioning/components/SelectedProjectBanner';
 import {ConsoleUserProjectWithExtension} from '../../CustomerDashboard/pages/Apps/App/CloudProvisioning/pages/CloudProvisioningOutlet';
-import {ProductCardRevamp} from '../../GetApp/components/ProductCard/ProductCard';
+import {ProductCardRevamp} from '../../ProductPurchase/components/ProductCard/ProductCard';
 import {useOAuth2OutletContext} from '../OAuth2AuthorizeOutlet';
 
 const EnvironmentSelection = () => {
