+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/ProductPurchase/pages/App/License/index.tsx
@@ -1,4 +1,4 @@
-/**
+	/**
  * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
@@ -21,7 +21,7 @@ import {PaymentMethodType} from '../../../types';
 import PaidLicense from './PaidLicense';
 import TrialLicense from './TrialLicense';
 
-import '../../../../GetApp/styles/index.scss';
+import '../../../../ProductPurchase/components/StepWizard/StepWizard.scss';
 import {cartStore} from '../../../store/CartStore';
 
 const isContinueButtonDisabled = () => {
