+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/utils/getIcon.ts
@@ -1,0 +1,11 @@
+import {getThumbnailByProductAttachment} from './util';
+
+const getIconUrl = (product?: DeliveryProduct) => {
+	const iconURL = product
+		? getThumbnailByProductAttachment(product.images)?.split('/o/')
+		: '';
+
+	return iconURL ? `/o/${iconURL[1]}` : '';
+};
+
+export {getIconUrl};
