+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/utils/getProductOrderTypes.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {ProductType} from '../../../enums/Product';
+import {ProductType} from '../enums/Product';
 
 const productTypeERC = {
 	[ProductType.CLIENT_EXTENSION]: 'CLIENT_EXTENSION',
